+++ packages/data/src/mso/index.ts
@@ -1,14 +1,14 @@
 import * as MSO from '../types/mso';
-import { operators, versions } from './conditionals';
 import { alternatives } from './alternatives';
 import { standard } from './standard';
 
-export * from './conditionals';
+export * from './operators';
+export * from './versions';
 export * from './features';
 export * from './syntaxes';
 export * from './types';
 
-export { alternatives, standard, operators, versions, MSO };
+export { alternatives, standard, MSO };
 
 export const properties = {
   ...alternatives,
