+++ packages/data/src/mso/versions.ts
@@ -1,0 +1,40 @@
+import { Versions } from '../types/mso';
+
+export const versions: Versions = {
+  ie: {
+    syntax: 'IE',
+    versions: ['2000', '2002', '2003'],
+  },
+  mso: {
+    syntax: 'mso',
+    versions: ['2000', '2002', '2003', '2007', '2010', '2013', '2016'],
+  },
+  mso9: {
+    syntax: 'mso 9',
+    versions: ['2000'],
+  },
+  mso10: {
+    syntax: 'mso 10',
+    versions: ['2002'],
+  },
+  mso11: {
+    syntax: 'mso 11',
+    versions: ['2003'],
+  },
+  mso12: {
+    syntax: 'mso 12',
+    versions: ['2007'],
+  },
+  mso14: {
+    syntax: 'mso 14',
+    versions: ['2010'],
+  },
+  mso15: {
+    syntax: 'mso 15',
+    versions: ['2013'],
+  },
+  mso16: {
+    syntax: 'mso 16',
+    versions: ['2016'],
+  },
+};
